+++ src/input_context/action_binding.rs
@@ -287,7 +287,7 @@ impl ActionBinding {
         self
     }
 
-    /// Type-erased version for [`Actions::mock`].
+    /// Type-erased version for [`UntypedActions::mock`].
     pub(super) fn mock(&mut self, state: ActionState, value: ActionValue, span: MockSpan) {
         debug!(
             "mocking `{}` with `{state:?}` and `{value:?}` for `{span:?}`",
@@ -303,11 +303,9 @@ impl ActionBinding {
 
     pub(super) fn update(
         &mut self,
-        commands: &mut Commands,
         reader: &mut InputReader,
         action_map: &mut TypeIdMap<UntypedAction>,
         time: &InputTime,
-        entity: Entity,
     ) {
         let (state, value) = self
             .update_from_mock(time.delta())
@@ -318,7 +316,6 @@ impl ActionBinding {
             .expect("actions and bindings should have matching type IDs");
 
         action.update(time, state, value);
-        action.trigger_events(commands, entity);
     }
 
     pub(super) fn update_from_reader(
@@ -418,6 +415,18 @@ impl ActionBinding {
         Some((state, value))
     }
 
+    pub(super) fn trigger(
+        &mut self,
+        commands: &mut Commands,
+        action_map: &mut TypeIdMap<UntypedAction>,
+        entity: Entity,
+    ) {
+        let action = action_map
+            .get_mut(&self.type_id)
+            .expect("actions and bindings should have matching type IDs");
+        action.trigger_events(commands, entity);
+    }
+
     pub(super) fn type_id(&self) -> TypeId {
         self.type_id
     }
@@ -448,7 +457,7 @@ struct ActionMock {
 
 /// Specifies how long a mock input should remain active.
 ///
-/// See also [`Actions::mock`].
+/// See also [`UntypedActions::mock`].
 #[derive(Clone, Copy, Debug)]
 pub enum MockSpan {
     /// Mock for a fixed number of context evaluations.