+++ src/input_reader.rs
@@ -27,9 +27,9 @@ pub(crate) struct InputReader<'w, 's> {
 }
 
 impl InputReader<'_, '_> {
-    /// Resets all consumed values and reads mouse events.
+    /// Clears all consumed values and updates reset input.
     pub(crate) fn update_state(&mut self) {
-        self.consumed.reset();
+        self.consumed.clear();
 
         // Temporary take the original value to avoid issues with the borrow checker.
         let mut reset_input = core::mem::take(&mut *self.reset_input);
@@ -224,7 +224,7 @@ impl InputReader<'_, '_> {
 /// use bevy_enhanced_input::prelude::*;
 ///
 /// # let mut app = App::new();
-/// app.add_systems(PreUpdate, disable_mouse.before(EnhancedInputSystem));
+/// app.add_systems(PreUpdate, disable_mouse.before(EnhancedInputSet::Update));
 ///
 /// fn disable_mouse(
 ///     mut action_sources: ResMut<ActionSources>,
@@ -271,7 +271,7 @@ struct ConsumedInput {
 }
 
 impl ConsumedInput {
-    fn reset(&mut self) {
+    fn clear(&mut self) {
         self.keys.clear();
         self.mod_keys = ModKeys::empty();
         self.mouse_buttons.clear();